requirements.txt
pandas
scipy
numpy
scikit-learn
http://download.pytorch.org/whl/cu75/torch-0.2.0.post3-cp36-cp36m-manylinux1_x86_64.whl
torchvision
xlrd
